+++ CHANGELOG.md
@@ -8,9 +8,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/bbx0/borgreport/commits/main/) - 2024-10-02
+## [Unreleased](https://github.com/bbx0/borgreport/compare/v0.1.0...HEAD) - 2024-10-12
 
-TBD
+### Changed
+
+- Use `--bypass-lock` for the `borg info` command
 
 ## [0.1.0](https://github.com/bbx0/borgreport/releases/tag/v0.1.0) - 2024-10-02